+++ node/packages/aws-lambda-sdk/test/fixtures/lambdas/esm-sdk/index.js
@@ -23,6 +23,9 @@ export const handler = async () => {
   sdk.setTag('user.tag', `example:${invocationId}`);
 
   console.warn('Consoled warning', 12, true);
+
+  sdk._createTraceSpan('custom.not.closed');
+
   return {
     name: sdk.name,
     version: sdk.version,