+++ python/packages/sdk-schema/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 0.2.1 (2023-04-07)
+
+### Features
+
+- AWS Lambda runtime related tags
+
 ### 0.2.0 (2023-03-28)
 
 ### ⚠ BREAKING CHANGES